Admin features + reviewed texts
The admin of the application have got the ability of changing the
interval between the update of the exchange rates, and the initial
capital given to the user who starts a new game.

Also, the database table called "Settings" has been modified, and now
also contains the "Interval" and "Initial Sum" columns. This change has
been reflected in the application.

The messages that are shown to the application users have been reviewed
and modified.